+++ app/uk/gov/hmrc/rdsdatacacheproxy/Module.scala
@@ -19,7 +19,7 @@ package uk.gov.hmrc.rdsdatacacheproxy
 import play.api.inject.{Binding, Module as AppModule}
 import play.api.{Configuration, Environment}
 import uk.gov.hmrc.rdsdatacacheproxy.actions.{AuthAction, DefaultAuthAction}
-import uk.gov.hmrc.rdsdatacacheproxy.connectors.{RdsDataSource, RdsDatacacheRepository, RdsStub}
+import uk.gov.hmrc.rdsdatacacheproxy.repositories.{RdsDataSource, RdsDatacacheRepository, RdsStub}
 import uk.gov.hmrc.rdsdatacacheproxy.controllers.DirectDebitController
 
 class Module extends AppModule:
