+++ app/uk/gov/hmrc/rdsdatacacheproxy/repositories/RdsStub.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.rdsdatacacheproxy.connectors
+package uk.gov.hmrc.rdsdatacacheproxy.repositories
 
 import uk.gov.hmrc.rdsdatacacheproxy.models.responses.{DDIReference, DirectDebit, EarliestPaymentDate, UserDebits}
 import uk.gov.hmrc.rdsdatacacheproxy.utils.StubUtils
@@ -33,7 +33,7 @@ class RdsStub @Inject()() extends RdsDataSource:
   //  val serviceUrl: String = servicesConfig.baseUrl("rds")
 
   // Remove this once real stubbing exists
-  private[connectors] val stubData = new StubUtils()
+  private[repositories] val stubData = new StubUtils()
 
   def getDirectDebits(id: String, start: Int, max: Int): Future[UserDebits] =
     val debits: Seq[DirectDebit] = for(i <- 1 to max) yield stubData.randomDirectDebit(i)
