+++ test/uk/gov/hmrc/rdsdatacacheproxy/repositories/RdsDatacacheRepositorySpec.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.rdsdatacacheproxy.connectors
+package uk.gov.hmrc.rdsdatacacheproxy.repositories
 
 import org.mockito.ArgumentMatchers.*
 import org.mockito.Mockito.*
