+++ test/uk/gov/hmrc/rdsdatacacheproxy/services/DirectDebitServiceSpec.scala
@@ -22,7 +22,7 @@ import org.scalatest.concurrent.{IntegrationPatience, ScalaFutures}
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
 import org.scalatestplus.mockito.MockitoSugar
-import uk.gov.hmrc.rdsdatacacheproxy.connectors.RdsStub
+import uk.gov.hmrc.rdsdatacacheproxy.repositories.RdsStub
 import uk.gov.hmrc.rdsdatacacheproxy.models.responses.{DDIReference, DirectDebit, EarliestPaymentDate, UserDebits}
 
 import java.time.{LocalDate, LocalDateTime}
